pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piclock3"
version = "0.1.0"
description = "Analogue clock face for Raspberry Pi Zero with circular display"
requires-python = ">=3.11"
dependencies = [
    "pycairo>=1.25",
    "pygame>=2.5",
    "flask>=3.0",
    "numpy>=1.26",
    "tzdata>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
svg = [
    "cairosvg>=2.7",
]

[project.scripts]
piclock = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
